Creates reminder script for Zoom calendar events
This uses the Google Calendar API to create a reminder script for each
of the events in your calendar that use Zoom (up to 5 at the moment).
When this script is run that it uses Terminal-Notifier to send a
notification to your desktop with the event title as the title and
when clicked opens the zoom link specified in the events location.